+++ pages/index.js
@@ -1,11 +1,7 @@
-import '../components/tap_events'
-
 import { Component } from 'react'
 import Head from 'next/head'
 import Layout from '../components/layout/Layout'
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import Navigation from '../components/navigation/Navigation'
-import getMuiTheme from 'material-ui/styles/getMuiTheme'
 
 export default class extends Component {
   static async getInitialProps({ req }) {
@@ -16,11 +12,9 @@ export default class extends Component {
 
   render() {
     return (
-      <MuiThemeProvider muiTheme={getMuiTheme({userAgent: this.props.userAgent})}>
-        <Layout>
-          <h1>Hello World</h1>
-        </Layout>
-      </MuiThemeProvider>
+      <Layout userAgent={this.props.userAgent}>
+        <h1>Hello World</h1>
+      </Layout>
     )
   }
 }
